#2276 - More Spring auto-configuration
- Allow using project service without project initializers
- Fix dependencies
